Extract movie insert to a show insert method
