+++ api/src/models/entry/index.ts
@@ -12,7 +12,7 @@ export const Entry = t.Union([Episode, MovieEntry, Special]);
 export type Entry = typeof Entry.static;
 
 export const SeedEntry = t.Union([SeedEpisode, SeedMovieEntry, SeedSpecial]);
-export type SeedEntry = typeof Entry.static;
+export type SeedEntry = typeof SeedEntry.static;
 
 export * from "./episode";
 export * from "./movie-entry";
